Move VK_ to KeyCode map to `keyCodes.ts`
